:recycle: refactor(cucumber): cucumber tests related to fee on one feature file, and in one step
