--- src/test/resources/features/exception_is_thrown_when_cep_is_null.feature
@@ -1,6 +1,0 @@
-Feature: Exception is thrown when cep null is sent in request
-
-Scenario: send an invalid cep
-    Given null cep
-    When request is made with null cep
-    Then RequestException is thrown explaining that field cep is mandatory
