--- src/test/resources/features/exception_is_thrown_when_invalid_cep_is_sent.feature
@@ -1,6 +1,0 @@
-Feature: Exception is thrown when invalid cep is sent
-
-Scenario: send an invalid cep
-    Given invalid cep
-    When request is made with invalid cep
-    Then RequestException is thrown explaining the error
