--- src/test/resources/features/receiving_feeResponse_when_a_valid_cep_is_given.feature
@@ -1,6 +1,0 @@
-Feature: Receiving feeResponse when a valid cep is given
-
-Scenario: send a valid cep
-    Given valid cep
-    When request is made
-    Then FeeResponse is returned according to cep
